backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
pyodbc
pydantic
